Add doc to UI and visualizations components. And refactor the name of some packages
