+++ app/src/main/java/chpc/prediction/WekaModule.java
@@ -1,7 +1,6 @@
-package chpc.weka;
+package chpc.prediction;
 
 import chpc.dataLoader.NHPIRecord;
-import chpc.visualizations.PopUpWindow;
 
 import java.util.ArrayList;
 import java.util.List;
