+++ app/src/main/java/chpc/visualizations/views/Bar.java
@@ -11,12 +11,19 @@ import chpc.dataLoader.NHPIRecord;
 
 import java.awt.*;
 import java.util.List;
-
+/**
+ * A class can create a window containing a bar chart
+ */
 public class Bar extends View {
   DefaultCategoryDataset dataset;
   String dataGroup;
   boolean predicted;
 
+  /**
+   * constructor to  create a window containing a bar chart
+   * @param dataGroup it specifies the which group of data is used to display on chart
+   * @param predicted it specifies if the chart is used to show predicated datq
+   */
   public Bar(String dataGroup, boolean predicted) {
     this.predicted = predicted;
     this.dataGroup = dataGroup;
@@ -25,6 +32,9 @@ public class Bar extends View {
     createChart();
   }
 
+  /**
+   * access selected data
+   */
   public void addDataset() {
 
     dataset = new DefaultCategoryDataset();
@@ -41,15 +51,20 @@ public class Bar extends View {
 
   }
 
-  public void addDataset(List<NHPIRecord> inputList) {
 
-    dataset = new DefaultCategoryDataset();
-    for (NHPIRecord r : inputList) {
-      dataset.setValue(r.getValue(), r.getGeo(), r.getRefDate());
-    }
+//  public void addDataset(List<NHPIRecord> inputList) {
+//
+//    dataset = new DefaultCategoryDataset();
+//    for (NHPIRecord r : inputList) {
+//      dataset.setValue(r.getValue(), r.getGeo(), r.getRefDate());
+//    }
+//
+//  }
 
-  }
-
+  /**
+   *
+   * @return a plotted a bar chart
+   */
   public JFreeChart plotView() {
     CategoryPlot plot = new CategoryPlot();
 
@@ -62,12 +77,9 @@ public class Bar extends View {
     plot.setDomainAxis(domainAxis);
     plot.setRangeAxis(new NumberAxis("NHPI"));
 
-    // plot.setDataset(1, dataset2);
     plot.setRenderer(1, barrenderer2);
-    // plot.setRangeAxis(1, new NumberAxis("US$"));
 
     plot.mapDatasetToRangeAxis(0, 0);// 1st dataset to 1st y-axis
-    // plot.mapDatasetToRangeAxis(1, 1); // 2nd dataset to 2nd y-axis
 
     JFreeChart barChart = new JFreeChart("Bar: New housing price index (NHPI) ",
         new Font("Serif", Font.BOLD, 18), plot, true);
